Active flag in /floor/get API and /factor/get
