+++ api/rest/factor_test.go
@@ -15,17 +15,13 @@ import (
 	"github.com/m6yf/bcwork/dto"
 	"github.com/m6yf/bcwork/models"
 	"github.com/m6yf/bcwork/utils/constant"
-	"github.com/m6yf/bcwork/validations"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
 	"github.com/volatiletech/null/v8"
 )
 
 func TestValidateFactors(t *testing.T) {
-	app := fiber.New()
-	app.Post("/factor", validations.ValidateFactor, func(c *fiber.Ctx) error {
-		return c.SendString("Factor created successfully")
-	})
+	endpoint := "/test/factor"
 
 	tests := []struct {
 		name         string
@@ -67,9 +63,9 @@ func TestValidateFactors(t *testing.T) {
 	}
 
 	for _, test := range tests {
-		req := httptest.NewRequest("POST", "/factor", strings.NewReader(test.body))
+		req := httptest.NewRequest(fiber.MethodPost, endpoint, strings.NewReader(test.body))
 		req.Header.Set("Content-Type", "application/json")
-		resp, err := app.Test(req)
+		resp, err := appTest.Test(req)
 		if err != nil {
 			t.Errorf("Test %s failed: %s", test.name, err)
 			continue
@@ -172,12 +168,12 @@ func TestCreateFactorMetadataGeneration(t *testing.T) {
 func Test_Factor_ToModel_(t *testing.T) {
 	tests := []struct {
 		name     string
-		factor   *core.Factor
+		factor   *dto.Factor
 		expected *models.Factor
 	}{
 		{
 			name: "All fields populated",
-			factor: &core.Factor{
+			factor: &dto.Factor{
 				RuleId:        "50afedac-d41a-53b0-a922-2c64c6e80623",
 				Publisher:     "Publisher1",
 				Domain:        "example.com",
@@ -202,7 +198,7 @@ func Test_Factor_ToModel_(t *testing.T) {
 		},
 		{
 			name: "Some fields empty",
-			factor: &core.Factor{
+			factor: &dto.Factor{
 				RuleId:        "d823a92a-83e5-5c2b-a067-b982d6cdfaf8",
 				Publisher:     "Publisher2",
 				Domain:        "example.org",
@@ -227,7 +223,7 @@ func Test_Factor_ToModel_(t *testing.T) {
 		},
 		{
 			name: "All fields empty",
-			factor: &core.Factor{
+			factor: &dto.Factor{
 				RuleId:        "966affd7-d087-57a2-baff-55b926f4c32d",
 				Publisher:     "",
 				Domain:        "",
@@ -264,7 +260,7 @@ func Test_Factor_ToModel(t *testing.T) {
 	t.Parallel()
 
 	type args struct {
-		factor *core.Factor
+		factor *dto.Factor
 	}
 
 	tests := []struct {
@@ -275,7 +271,7 @@ func Test_Factor_ToModel(t *testing.T) {
 		{
 			name: "All fields populated",
 			args: args{
-				factor: &core.Factor{
+				factor: &dto.Factor{
 					RuleId:        "50afedac-d41a-53b0-a922-2c64c6e80623",
 					Publisher:     "Publisher1",
 					Domain:        "example.com",
@@ -302,7 +298,7 @@ func Test_Factor_ToModel(t *testing.T) {
 		{
 			name: "Some fields empty",
 			args: args{
-				factor: &core.Factor{
+				factor: &dto.Factor{
 					RuleId:        "d823a92a-83e5-5c2b-a067-b982d6cdfaf8",
 					Publisher:     "Publisher2",
 					Domain:        "example.org",
@@ -329,7 +325,7 @@ func Test_Factor_ToModel(t *testing.T) {
 		{
 			name: "All fields empty",
 			args: args{
-				factor: &core.Factor{
+				factor: &dto.Factor{
 					RuleId:        "966affd7-d087-57a2-baff-55b926f4c32d",
 					Publisher:     "",
 					Domain:        "",
@@ -396,7 +392,7 @@ func TestFactorHistory(t *testing.T) {
 					Subject:      "Bidder Targeting",
 					Item:         "333_3.com_af_tablet_windowsphone_opera_rectangle",
 					Changes: []dto.Changes{
-						{Property: "factor", OldValue: nil, NewValue: float64(0.02)},
+						{Property: "factor", NewValue: float64(0.02), OldValue: nil},
 					},
 				},
 			},
