+++ core/floor.go
@@ -6,11 +6,10 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/m6yf/bcwork/dto"
 
 	"sort"
 	"strings"
-
-	"github.com/volatiletech/null/v8"
 
 	"github.com/m6yf/bcwork/bcdb"
 	"github.com/m6yf/bcwork/bcdb/filter"
@@ -20,8 +19,6 @@ import (
 	"github.com/m6yf/bcwork/models"
 	"github.com/m6yf/bcwork/modules/history"
 	"github.com/m6yf/bcwork/utils"
-	"github.com/m6yf/bcwork/utils/bcguid"
-	"github.com/m6yf/bcwork/utils/constant"
 	"github.com/rotisserie/eris"
 	"github.com/volatiletech/sqlboiler/v4/boil"
 	"github.com/volatiletech/sqlboiler/v4/queries/qm"
@@ -37,19 +34,6 @@ func NewFloorService(historyModule history.HistoryModule) *FloorService {
 	}
 }
 
-type Floor struct {
-	RuleId        string  `boil:"rule_id" json:"rule_id" toml:"rule_id" yaml:"rule_id"`
-	Publisher     string  `boil:"publisher" json:"publisher" toml:"publisher" yaml:"publisher"`
-	PublisherName string  `boil:"publisher_name" json:"publisher_name" toml:"publisher_name" yaml:"publisher_name"`
-	Domain        string  `boil:"domain" json:"domain" toml:"domain" yaml:"domain"`
-	Country       string  `boil:"country" json:"country" toml:"country" yaml:"country"`
-	Device        string  `boil:"device" json:"device" toml:"device" yaml:"device"`
-	Floor         float64 `boil:"floor" json:"floor" toml:"floor" yaml:"floor"`
-	Browser       string  `boil:"browser" json:"browser" toml:"browser" yaml:"browser"`
-	OS            string  `boil:"os" json:"os" toml:"os" yaml:"os"`
-	PlacementType string  `boil:"placement_type" json:"placement_type" toml:"placement_type" yaml:"placement_type"`
-}
-
 type GetFloorOptions struct {
 	Filter     FloorFilter            `json:"filter"`
 	Pagination *pagination.Pagination `json:"pagination"`
@@ -62,9 +46,8 @@ type FloorFilter struct {
 	Domain    filter.StringArrayFilter `json:"domain"`
 	Country   filter.StringArrayFilter `json:"country"`
 	Device    filter.StringArrayFilter `json:"device"`
+	Active    filter.StringArrayFilter `json:"active"`
 }
-
-type FloorSlice []*Floor
 
 type FloorRealtimeRecord struct {
 	Rule   string  `json:"rule"`
@@ -72,102 +55,7 @@ type FloorRealtimeRecord struct {
 	RuleID string  `json:"rule_id"`
 }
 
-func (floor *Floor) FromModel(mod *models.Floor) error {
-	floor.RuleId = mod.RuleID
-	floor.Publisher = mod.Publisher
-	floor.Domain = mod.Domain
-	floor.Floor = mod.Floor
-	floor.RuleId = mod.RuleID
-
-	if mod.R != nil && mod.R.FloorPublisher != nil {
-		floor.PublisherName = mod.R.FloorPublisher.Name
-	}
-
-	if mod.Os.Valid {
-		floor.OS = mod.Os.String
-	}
-
-	if mod.Country.Valid {
-		floor.Country = mod.Country.String
-	}
-
-	if mod.Device.Valid {
-		floor.Device = mod.Device.String
-	}
-
-	if mod.PlacementType.Valid {
-		floor.PlacementType = mod.PlacementType.String
-	}
-
-	if mod.Browser.Valid {
-		floor.Browser = mod.Browser.String
-	}
-
-	return nil
-}
-
-func (floor *Floor) GetRuleID() string {
-	if len(floor.RuleId) > 0 {
-		return floor.RuleId
-	} else {
-		return bcguid.NewFrom(floor.GetFormula())
-	}
-}
-
-func (floor *Floor) GetFormula() string {
-	p := floor.Publisher
-	if p == "" {
-		p = "*"
-	}
-
-	d := floor.Domain
-	if d == "" {
-		d = "*"
-	}
-
-	c := floor.Country
-	if c == "" {
-		c = "*"
-	}
-
-	os := floor.OS
-	if os == "" {
-		os = "*"
-	}
-
-	dt := floor.Device
-	if dt == "" {
-		dt = "*"
-	}
-
-	pt := floor.PlacementType
-	if pt == "" {
-		pt = "*"
-	}
-
-	b := floor.Browser
-	if b == "" {
-		b = "*"
-	}
-
-	return fmt.Sprintf("p=%s__d=%s__c=%s__os=%s__dt=%s__pt=%s__b=%s", p, d, c, os, dt, pt, b)
-
-}
-
-func (cs *FloorSlice) FromModel(slice models.FloorSlice) error {
-	for _, mod := range slice {
-		c := Floor{}
-		err := c.FromModel(mod)
-		if err != nil {
-			return eris.Cause(err)
-		}
-		*cs = append(*cs, &c)
-	}
-
-	return nil
-}
-
-func (f *FloorService) GetFloors(ctx context.Context, ops GetFloorOptions) (FloorSlice, error) {
+func (f *FloorService) GetFloors(ctx context.Context, ops GetFloorOptions) (dto.FloorSlice, error) {
 	qmods := ops.Filter.QueryMod().
 		Order(ops.Order, nil, models.FloorColumns.Publisher).
 		AddArray(ops.Pagination.Do()).
@@ -180,7 +68,7 @@ func (f *FloorService) GetFloors(ctx context.Context, ops GetFloorOptions) (Floo
 		return nil, eris.Wrap(err, "Failed to retrieve floors")
 	}
 
-	res := make(FloorSlice, 0)
+	res := make(dto.FloorSlice, 0)
 	err = res.FromModel(mods)
 	if err != nil {
 		return nil, err
@@ -213,10 +101,14 @@ func (filter *FloorFilter) QueryMod() qmods.QueryModsSlice {
 		mods = append(mods, filter.Country.AndIn(models.FloorColumns.Country))
 	}
 
+	if len(filter.Active) > 0 {
+		mods = append(mods, filter.Active.AndIn(models.FloorColumns.Active))
+	}
+
 	return mods
 }
 
-func (f *FloorService) UpdateFloorMetaData(ctx context.Context, data constant.FloorUpdateRequest) error {
+func (f *FloorService) UpdateFloorMetaData(ctx context.Context, data dto.FloorUpdateRequest) error {
 	_, err := json.Marshal(data)
 	if err != nil {
 		return fmt.Errorf("failed to parse hash value for floor: %w", err)
@@ -230,51 +122,10 @@ func (f *FloorService) UpdateFloorMetaData(ctx context.Context, data constant.Fl
 	return nil
 }
 
-func (floor *Floor) ToModel() *models.Floor {
-	mod := models.Floor{
-		Floor:     floor.Floor,
-		Publisher: floor.Publisher,
-		Domain:    floor.Domain,
-		RuleID:    floor.RuleId,
-	}
-
-	if floor.Country != "" {
-		mod.Country = null.StringFrom(floor.Country)
-	} else {
-		mod.Country = null.String{}
-	}
-
-	if floor.OS != "" {
-		mod.Os = null.StringFrom(floor.OS)
-	} else {
-		mod.Os = null.String{}
-	}
-
-	if floor.Device != "" {
-		mod.Device = null.StringFrom(floor.Device)
-	} else {
-		mod.Device = null.String{}
-	}
-
-	if floor.PlacementType != "" {
-		mod.PlacementType = null.StringFrom(floor.PlacementType)
-	} else {
-		mod.PlacementType = null.String{}
-	}
-
-	if floor.Browser != "" {
-		mod.Browser = null.StringFrom(floor.Browser)
-	} else {
-		mod.Browser = null.String{}
-	}
-
-	return &mod
-}
-
-func (f *FloorService) UpdateFloors(ctx context.Context, data constant.FloorUpdateRequest) (bool, error) {
+func (f *FloorService) UpdateFloors(ctx context.Context, data dto.FloorUpdateRequest) (bool, error) {
 	var isInsert bool
 
-	floor := Floor{
+	floor := dto.Floor{
 		Publisher:     data.Publisher,
 		Domain:        data.Domain,
 		Country:       data.Country,
@@ -284,7 +135,9 @@ func (f *FloorService) UpdateFloors(ctx context.Context, data constant.FloorUpda
 		OS:            data.OS,
 		PlacementType: data.PlacementType,
 		RuleId:        data.RuleId,
+		Active:        data.Active,
 	}
+
 	if len(floor.RuleId) == 0 {
 		floor.RuleId = floor.GetRuleID()
 	}
@@ -294,6 +147,7 @@ func (f *FloorService) UpdateFloors(ctx context.Context, data constant.FloorUpda
 	oldMod, err := models.Floors(
 		models.FloorWhere.RuleID.EQ(mod.RuleID),
 	).One(ctx, bcdb.DB())
+
 	if err != nil && err != sql.ErrNoRows {
 		return false, err
 	}
@@ -321,8 +175,7 @@ func (f *FloorService) UpdateFloors(ctx context.Context, data constant.FloorUpda
 	return isInsert, nil
 }
 
-func sendFloorToRT(ctx context.Context, updateRequest constant.FloorUpdateRequest) error {
-	const PREFIX string = "price:floor:v2"
+func sendFloorToRT(ctx context.Context, updateRequest dto.FloorUpdateRequest) error {
 	modFloor, err := FloorQuery(ctx, updateRequest)
 
 	if err != nil && err != sql.ErrNoRows {
@@ -343,7 +196,7 @@ func sendFloorToRT(ctx context.Context, updateRequest constant.FloorUpdateReques
 	}
 
 	key := utils.GetMetadataObject(updateRequest)
-	metadataKey := utils.CreateMetadataKey(key, PREFIX)
+	metadataKey := utils.CreateMetadataKey(key, utils.FloorMetaDataKeyPrefix)
 	metadataValue := utils.CreateMetadataObject(updateRequest, metadataKey, value)
 
 	err = metadataValue.Insert(ctx, bcdb.DB(), boil.Infer())
@@ -354,7 +207,7 @@ func sendFloorToRT(ctx context.Context, updateRequest constant.FloorUpdateReques
 	return nil
 }
 
-func FloorQuery(ctx context.Context, updateRequest constant.FloorUpdateRequest) (models.FloorSlice, error) {
+func FloorQuery(ctx context.Context, updateRequest dto.FloorUpdateRequest) (models.FloorSlice, error) {
 	modFloor, err := models.Floors(
 		models.FloorWhere.Domain.EQ(updateRequest.Domain),
 		models.FloorWhere.Publisher.EQ(updateRequest.Publisher),
@@ -365,7 +218,7 @@ func FloorQuery(ctx context.Context, updateRequest constant.FloorUpdateRequest) 
 
 func CreateFloorMetadata(modFloor models.FloorSlice, finalRules []FloorRealtimeRecord) []FloorRealtimeRecord {
 	if len(modFloor) != 0 {
-		floors := make(FloorSlice, 0)
+		floors := make(dto.FloorSlice, 0)
 		floors.FromModel(modFloor)
 
 		for _, floor := range floors {
